creating the moment variable
