+++ package.json
@@ -17,7 +17,6 @@
     "normalize.css": "^8.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-moment": "^1.1.3",
     "sass": "^1.72.0"
   },
   "devDependencies": {
