+++ src/App.tsx
@@ -1,18 +1,11 @@
 import Clock from './components/Clock/Clock';
 import styles from './App.module.scss';
-import targetTime from './utils';
+// import targetTime from './utils';
 
 function App() {
-  const getClientUrl = () => {
-    const urlCurrent = window.location.href;
-    const clientTtimeStamp = new Date().getTime();
-    const clientTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-    const urlClient = `${urlCurrent}/?ts=${clientTtimeStamp}&tz=${clientTimeZone}`;
-    return urlClient;
-  };
+  // const clientUrl = 'http://localhost:5173/?ts=1618598453&tz=Europa/Moscow';
 
-  // console.log(`Ссылка: ${urlCurrent}/?ts=${clientTtimeStamp}tz=${clientTimeZone} `);
-  // const url = 'https://timesync.club/?ts=1712745573031&tz=Asia/Tokyo';
+  // const today = moment().format('dddd, MMMM Do YYYY, h:mm:ss a');
 
   return (
     <>
@@ -21,7 +14,7 @@ function App() {
           Secret meeting of the <span> Masons</span>
         </h1>
         <div className={styles.block}>
-          <Clock time={new Date(targetTime(getClientUrl()))} />
+          <Clock time={new Date() } />
           <Clock time={new Date()} />
         </div>
       </div>
