+++ src/components/Clock/Clock.module.scss
@@ -15,7 +15,7 @@
     width: 14px;
     height: 14px;
     position: absolute;
-    background: #f57474;
+    background: #171616;
     border-radius: 30%;
     z-index: 10;
     top: calc(50% - 7px);
@@ -35,7 +35,7 @@
       content: '';
       width: 8px;
       height: 75px;
-      background-color: #302e2e;
+      background-color: #171616;
       border-radius: 15px 15px 0 0;
     }
   }
@@ -53,7 +53,7 @@
       content: '';
       width: 4px;
       height: 100px;
-      background-color: #e72424;
+      background-color: #171616;
       border-radius: 4px 4px 0 0;
     }
   }
